Separate out the testing compatibility file.
- Testing compatibility includes dependencies not present in base images.
- Includes a fix for usermod ordering needed by older distros.
