+++ google_compute_engine/accounts/tests/accounts_daemon_test.py
@@ -18,8 +18,8 @@
 import datetime
 
 from google_compute_engine.accounts import accounts_daemon
-from google_compute_engine.compat import mock
-from google_compute_engine.compat import unittest
+from google_compute_engine.test_compat import mock
+from google_compute_engine.test_compat import unittest
 
 
 class AccountsDaemonTest(unittest.TestCase):
