+++ google_compute_engine/clock_skew/tests/clock_skew_daemon_test.py
@@ -18,8 +18,8 @@
 import subprocess
 
 from google_compute_engine.clock_skew import clock_skew_daemon
-from google_compute_engine.compat import mock
-from google_compute_engine.compat import unittest
+from google_compute_engine.test_compat import mock
+from google_compute_engine.test_compat import unittest
 
 
 class ClockSkewDaemonTest(unittest.TestCase):
