+++ google_compute_engine/instance_setup/tests/instance_setup_test.py
@@ -17,9 +17,9 @@
 
 import subprocess
 
-from google_compute_engine.compat import mock
-from google_compute_engine.compat import unittest
 from google_compute_engine.instance_setup import instance_setup
+from google_compute_engine.test_compat import mock
+from google_compute_engine.test_compat import unittest
 
 
 class InstanceSetupTest(unittest.TestCase):
