+++ google_compute_engine/ip_forwarding/tests/ip_forwarding_utils_test.py
@@ -15,9 +15,9 @@
 
 """Unittest for ip_forwarding_utils.py module."""
 
-from google_compute_engine.compat import mock
-from google_compute_engine.compat import unittest
 from google_compute_engine.ip_forwarding import ip_forwarding_utils
+from google_compute_engine.test_compat import mock
+from google_compute_engine.test_compat import unittest
 
 
 class IpForwardingUtilsTest(unittest.TestCase):
