+++ google_compute_engine/metadata_scripts/tests/script_retriever_test.py
@@ -17,9 +17,9 @@
 
 import subprocess
 
-from google_compute_engine.compat import mock
-from google_compute_engine.compat import unittest
 from google_compute_engine.metadata_scripts import script_retriever
+from google_compute_engine.test_compat import mock
+from google_compute_engine.test_compat import unittest
 
 
 class ScriptRetrieverTest(unittest.TestCase):
