+++ google_compute_engine/tests/config_manager_test.py
@@ -16,9 +16,9 @@
 """Unittest for config_manager.py module."""
 
 from google_compute_engine import config_manager
-from google_compute_engine.compat import builtin
-from google_compute_engine.compat import mock
-from google_compute_engine.compat import unittest
+from google_compute_engine.test_compat import builtin
+from google_compute_engine.test_compat import mock
+from google_compute_engine.test_compat import unittest
 
 
 class ConfigManagerTest(unittest.TestCase):
