+++ google_compute_engine/tests/lock_file_test.py
@@ -16,8 +16,8 @@
 """Unittest for lock_file_test.py module."""
 
 from google_compute_engine import lock_file
-from google_compute_engine.compat import mock
-from google_compute_engine.compat import unittest
+from google_compute_engine.test_compat import mock
+from google_compute_engine.test_compat import unittest
 
 
 class LockFileTest(unittest.TestCase):
