+++ google_compute_engine/tests/logger_test.py
@@ -16,8 +16,8 @@
 """Unittest for logger.py module."""
 
 from google_compute_engine import logger
-from google_compute_engine.compat import mock
-from google_compute_engine.compat import unittest
+from google_compute_engine.test_compat import mock
+from google_compute_engine.test_compat import unittest
 
 
 class LoggerTest(unittest.TestCase):
