+++ google_compute_engine/tests/metadata_watcher_test.py
@@ -18,8 +18,8 @@
 import os
 
 from google_compute_engine import metadata_watcher
-from google_compute_engine.compat import mock
-from google_compute_engine.compat import unittest
+from google_compute_engine.test_compat import mock
+from google_compute_engine.test_compat import unittest
 
 
 class MetadataWatcherTest(unittest.TestCase):
